fix(utils): rollback fromat amount
