+++ packages/utils/src/format-amount/index.ts
